provider_api: move poller code out of the command
So it will be easier to unit test. Some day.
